More type changes

+++ cpp/include/cuvs/neighbors/ivf_pq.hpp
@@ -2866,7 +2866,7 @@ void make_rotation_matrix(raft::resources const& res,
  * @code{.cpp}
  *   raft::resources res;
  *   // allocate the buffer for the input centers
- *   auto cluster_centers = raft::make_device_matrix<float, uint32_t>(res, index.n_lists(),
+ *   auto cluster_centers = raft::make_device_matrix<float, int64_t>(res, index.n_lists(),
  index.dim());
  *   ... prepare ivf centroids in cluster_centers ...
  *   // reset the index
@@ -2886,7 +2886,7 @@ void make_rotation_matrix(raft::resources const& res,
  */
 void set_centers(raft::resources const& res,
                  index<int64_t>* index,
-                 raft::device_matrix_view<const float, uint32_t> cluster_centers);
+                 raft::device_matrix_view<const float, int64_t> cluster_centers);
 
 /**
  * @brief Public helper API for fetching a trained index's IVF centroids
